+++ dummyDatas/shopItems.json
@@ -8,9 +8,7 @@
       "score": 5,
       "reviewCount": 15,
       "img": "https://shop-phinf.pstatic.net/20210622_238/1624341463146EL2rM_JPEG/25477358831686602_1117910248.jpg?type=f296_296"
-    },
-
-    
+    },   
     {
       "id": 2,
       "name": "A4 클립보드 서류고정 클립파일 메뉴판 - You Make Me Smile",
