+++ src/pages/CartPage.jsx
@@ -1,11 +1,9 @@
 import { Title } from '../styles/Page';
 
-function CartPage() {
+export default function CartPage() {
   return (
     <div>
       <Title> cart </Title>
     </div>
   );
 }
-
-export default CartPage;
