+++ src/pages/NotFoundPage.jsx
@@ -1,7 +1,5 @@
 // TODO: 잘못된 경로인 경우 이전페이지로 이동하도록 변경하기
 
-function NotFoundPage() {
+export default function NotFoundPage() {
   return <p>404 Not Found</p>;
 }
-
-export default NotFoundPage;
