+++ src/pages/WorksPage.jsx
@@ -1,11 +1,9 @@
 import { Title } from '../styles/Page';
 
-function WorksPage() {
+export default function WorksPage() {
   return (
     <div>
       <Title> works </Title>
     </div>
   );
 }
-
-export default WorksPage;
